Split AggregatePath in interface and implementation.
This forces us to make some methods public, that have been private before.
