+++ spring-data-relational/src/main/java/org/springframework/data/relational/core/mapping/AggregatePath.java
@@ -18,12 +18,8 @@ package org.springframework.data.relational.core.mapping;
 
 import org.springframework.data.mapping.PersistentPropertyPath;
 import org.springframework.data.relational.core.sql.SqlIdentifier;
-import org.springframework.data.util.Lazy;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
-import org.springframework.util.StringUtils;
-
-import java.util.Objects;
 
 /**
  * Represents a path within an aggregate starting from the aggregate root.
@@ -31,57 +27,20 @@ import java.util.Objects;
  * @since 3.2
  * @author Jens Schauder
  */
-public class AggregatePath {
+public interface AggregatePath {
 
-	private final RelationalMappingContext context;
-
-	@Nullable private final RelationalPersistentEntity<?> rootType;
-
-	@Nullable private final PersistentPropertyPath<? extends RelationalPersistentProperty> path;
-	private final Lazy<SqlIdentifier> columnAlias = Lazy.of(() -> prefixWithTableAlias(getColumnName()));
-
-	public AggregatePath(RelationalMappingContext context,
-			PersistentPropertyPath<? extends RelationalPersistentProperty> path) {
-
-		Assert.notNull(context, "context must not be null");
-		Assert.notNull(path, "path must not be null");
-
-		this.context = context;
-		this.path = path;
-
-		this.rootType = null;
-	}
-
-	public AggregatePath(RelationalMappingContext context, RelationalPersistentEntity<?> rootType) {
-
-		Assert.notNull(context, "context must not be null");
-		Assert.notNull(rootType, "rootType must not be null");
-
-		this.context = context;
-		this.rootType = rootType;
-
-		this.path = null;
-	}
-
-	public static boolean isWritable(@Nullable PersistentPropertyPath<? extends RelationalPersistentProperty> path) {
+	static boolean isWritable(@Nullable PersistentPropertyPath<? extends RelationalPersistentProperty> path) {
 		return path == null || path.getLeafProperty().isWritable() && isWritable(path.getParentPath());
 	}
 
-	public boolean isRoot() {
-		return path == null;
-	}
+	boolean isRoot();
 
 	/**
 	 * The name of the column used to reference the id in the parent table.
 	 *
 	 * @throws IllegalStateException when called on an empty path.
 	 */
-	public SqlIdentifier getReverseColumnName() {
-
-		Assert.state(!isRoot(), "Empty paths don't have a reverse column name");
-
-		return path.getLeafProperty().getReverseColumnName(this);
-	}
+	SqlIdentifier getReverseColumnName();
 
 	/**
 	 * Returns the path that has the same beginning but is one segment shorter than this path.
@@ -89,18 +48,7 @@ public class AggregatePath {
 	 * @return the parent path. Guaranteed to be not {@literal null}.
 	 * @throws IllegalStateException when called on an empty path.
 	 */
-	public AggregatePath getParentPath() {
-
-		if (isRoot()) {
-			throw new IllegalStateException("The parent path of a root path is not defined.");
-		}
-
-		if (path.getLength() == 1) {
-			return context.getAggregatePath(path.getLeafProperty().getOwner());
-		}
-
-		return context.getAggregatePath(path.getParentPath());
-	}
+	AggregatePath getParentPath();
 
 	/**
 	 * The {@link RelationalPersistentEntity} associated with the leaf of this path.
@@ -108,9 +56,7 @@ public class AggregatePath {
 	 * @return Might return {@literal null} when called on a path that does not represent an entity.
 	 */
 	@Nullable
-	public RelationalPersistentEntity<?> getLeafEntity() {
-		return isRoot() ? rootType : context.getPersistentEntity(path.getLeafProperty().getActualType());
-	}
+	RelationalPersistentEntity<?> getLeafEntity();
 
 	/**
 	 * The {@link RelationalPersistentEntity} associated with the leaf of this path or throw {@link IllegalStateException}
@@ -119,56 +65,23 @@ public class AggregatePath {
 	 * @return the required {@link RelationalPersistentEntity} associated with the leaf of this path.
 	 * @throws IllegalStateException if the persistent entity cannot be resolved.
 	 */
-	public RelationalPersistentEntity<?> getRequiredLeafEntity() {
-
-		RelationalPersistentEntity<?> entity = getLeafEntity();
-
-		if (entity == null) {
-
-			throw new IllegalStateException(
-					String.format("Couldn't resolve leaf PersistentEntity for type %s", path.getLeafProperty().getActualType()));
-		}
-
-		return entity;
-	}
+	RelationalPersistentEntity<?> getRequiredLeafEntity();
 
 	/**
 	 * @return {@literal true} if this path represents an entity which has an Id attribute.
 	 */
-	public boolean hasIdProperty() {
-
-		RelationalPersistentEntity<?> leafEntity = getLeafEntity();
-		return leafEntity != null && leafEntity.hasIdProperty();
-	}
+	boolean hasIdProperty();
 
 	/**
 	 * Returns the longest ancestor path that has an {@link org.springframework.data.annotation.Id} property.
 	 *
 	 * @return A path that starts just as this path but is shorter. Guaranteed to be not {@literal null}.
 	 */
-	public AggregatePath getIdDefiningParentPath() {
+	AggregatePath getIdDefiningParentPath();
 
-		AggregatePath parent = getParentPath();
+	RelationalPersistentProperty getRequiredIdProperty();
 
-		if (isRoot()) {
-			return parent;
-		}
-
-		if (!parent.hasIdProperty()) {
-			return parent.getIdDefiningParentPath();
-		}
-
-		return parent;
-	}
-
-	public RelationalPersistentProperty getRequiredIdProperty() {
-		return isRoot() ? rootType.getRequiredIdProperty() : getRequiredLeafEntity().getRequiredIdProperty();
-
-	}
-
-	public int getLength() {
-		return isRoot() ? 0 : path.getLength();
-	}
+	int getLength();
 
 	/**
 	 * The column name used for the list index or map key of the leaf property of this path.
@@ -176,9 +89,7 @@ public class AggregatePath {
 	 * @return May be {@literal null}.
 	 */
 	@Nullable
-	public SqlIdentifier getQualifierColumn() {
-		return isRoot() ? null : path.getLeafProperty().getKeyColumn();
-	}
+	SqlIdentifier getQualifierColumn();
 
 	/**
 	 * The type of the qualifier column of the leaf property of this path or {@literal null} if this is not applicable.
@@ -186,74 +97,19 @@ public class AggregatePath {
 	 * @return may be {@literal null}.
 	 */
 	@Nullable
-	public Class<?> getQualifierColumnType() {
-
-		if (isRoot()) {
-			return null;
-		}
-		if (!path.getLeafProperty().isQualified()) {
-			return null;
-		}
-		return path.getLeafProperty().getQualifierColumnType();
-	}
+	Class<?> getQualifierColumnType();
 
 	/**
 	 * Returns {@literal true} exactly when the path is non empty and the leaf property an embedded one.
 	 *
 	 * @return if the leaf property is embedded.
 	 */
-	public boolean isEmbedded() {
-		return !isRoot() && path.getLeafProperty().isEmbedded();
-	}
+	boolean isEmbedded();
 
 	/**
 	 * @return {@literal true} when this is an empty path or the path references an entity.
 	 */
-	public boolean isEntity() {
-		return isRoot() || path.getLeafProperty().isEntity();
-	}
-
-	/**
-	 * Finds and returns the longest path with ich identical or an ancestor to the current path and maps directly to a
-	 * table.
-	 *
-	 * @return a path. Guaranteed to be not {@literal null}.
-	 */
-	private AggregatePath getTableOwningAncestor() {
-
-		return isEntity() && !isEmbedded() ? this : getParentPath().getTableOwningAncestor();
-	}
-
-	@Nullable
-	private SqlIdentifier assembleTableAlias() {
-
-		Assert.state(!isRoot(), "Path is null");
-
-		RelationalPersistentProperty leafProperty = path.getLeafProperty();
-		String prefix;
-		if (isEmbedded()) {
-			prefix = leafProperty.getEmbeddedPrefix();
-
-		} else {
-			prefix = leafProperty.getName();
-		}
-
-		if (path.getLength() == 1) {
-			Assert.notNull(prefix, "Prefix mus not be null");
-			return StringUtils.hasText(prefix) ? SqlIdentifier.quoted(prefix) : null;
-		}
-
-		AggregatePath parentPath = getParentPath();
-		SqlIdentifier sqlIdentifier = parentPath.assembleTableAlias();
-
-		if (sqlIdentifier != null) {
-
-			return parentPath.isEmbedded() ? sqlIdentifier.transform(name -> name.concat(prefix))
-					: sqlIdentifier.transform(name -> name + "_" + prefix);
-		}
-		return SqlIdentifier.quoted(prefix);
-
-	}
+	boolean isEntity();
 
 	/**
 	 * The alias used for the table on which this path is based.
@@ -261,13 +117,7 @@ public class AggregatePath {
 	 * @return a table alias, {@literal null} if the table owning path is the empty path.
 	 */
 	@Nullable
-	public SqlIdentifier getTableAlias() {
-
-		AggregatePath tableOwner = getTableOwningAncestor();
-
-		return tableOwner.isRoot() ? null : tableOwner.assembleTableAlias();
-
-	}
+	SqlIdentifier getTableAlias();
 
 	/**
 	 * The fully qualified name of the table this path is tied to or of the longest ancestor path that is actually tied to
@@ -276,77 +126,30 @@ public class AggregatePath {
 	 * @return the name of the table. Guaranteed to be not {@literal null}.
 	 * @since 3.0
 	 */
-	public SqlIdentifier getQualifiedTableName() {
-		return getTableOwningAncestor().getRequiredLeafEntity().getQualifiedTableName();
-	}
+	SqlIdentifier getQualifiedTableName();
 
 	/**
 	 * The name of the column used to represent this property in the database.
 	 *
 	 * @throws IllegalStateException when called on an empty path.
 	 */
-	public SqlIdentifier getColumnName() {
-
-		Assert.state(!isRoot(), "Path is null");
-
-		return assembleColumnName(path.getLeafProperty().getColumnName());
-	}
+	SqlIdentifier getColumnName();
 
 	/**
 	 * The alias for the column used to represent this property in the database.
 	 *
 	 * @throws IllegalStateException when called on an empty path.
 	 */
-	public SqlIdentifier getColumnAlias() {
-		return columnAlias.get();
-	}
+	SqlIdentifier getColumnAlias();
 
 	/**
 	 * The alias used in select for the column used to reference the id in the parent table.
 	 *
 	 * @throws IllegalStateException when called on an empty path.
 	 */
-	public SqlIdentifier getReverseColumnNameAlias() {
+	SqlIdentifier getReverseColumnNameAlias();
 
-		return prefixWithTableAlias(getReverseColumnName());
-	}
-
-	@Override
-	public String toString() {
-		return "AggregatePath["
-				+ (rootType == null ? path.getBaseProperty().getOwner().getType().getName() : rootType.getName()) + "]"
-				+ ((isRoot()) ? "/" : path.toDotPath());
-	}
-
-	private SqlIdentifier assembleColumnName(SqlIdentifier suffix) {
-
-		Assert.state(!isRoot(), "Path is null");
-
-		if (path.getLength() <= 1) {
-			return suffix;
-		}
-
-		PersistentPropertyPath<? extends RelationalPersistentProperty> parentPath = path.getParentPath();
-		RelationalPersistentProperty parentLeaf = parentPath.getLeafProperty();
-
-		if (!parentLeaf.isEmbedded()) {
-			return suffix;
-		}
-
-		String embeddedPrefix = parentLeaf.getEmbeddedPrefix();
-
-		return getParentPath().assembleColumnName(suffix.transform(embeddedPrefix::concat));
-	}
-
-	private SqlIdentifier prefixWithTableAlias(SqlIdentifier columnName) {
-
-		SqlIdentifier tableAlias = getTableAlias();
-		return tableAlias == null ? columnName : columnName.transform(name -> tableAlias.getReference() + "_" + name);
-	}
-
-	public String toDotPath() {
-		return isRoot() ? "" : path.toDotPath();
-	}
+	String toDotPath();
 
 	/**
 	 * Returns {@literal true} if there are multiple values for this path, i.e. if the path contains at least one element
@@ -354,67 +157,38 @@ public class AggregatePath {
 	 *
 	 * @return {@literal true} if the path contains a multivalued element.
 	 */
-	public boolean isMultiValued() {
-
-		return !isRoot() && //
-				(path.getLeafProperty().isCollectionLike() //
-						|| path.getLeafProperty().isQualified() //
-						|| getParentPath().isMultiValued() //
-				);
-	}
+	boolean isMultiValued();
 
 	/**
 	 * @return {@literal true} if the leaf property of this path is a {@link java.util.Map}.
 	 * @see RelationalPersistentProperty#isMap()
 	 */
-	public boolean isMap() {
-		return !isRoot() && path.getLeafProperty().isMap();
-	}
+	boolean isMap();
 
 	/**
 	 * @return {@literal true} when this is references a {@link java.util.List} or {@link java.util.Map}.
 	 */
-	public boolean isQualified() {
-		return !isRoot() && path.getLeafProperty().isQualified();
-	}
+	boolean isQualified();
 
-	public RelationalPersistentProperty getRequiredLeafProperty() {
+	RelationalPersistentProperty getRequiredLeafProperty();
 
-		if (isRoot()) {
-			throw new IllegalStateException("root path does not have a leaf property");
-		}
-		return path.getLeafProperty();
-	}
-
-	public RelationalPersistentProperty getBaseProperty() {
-
-		if (isRoot()) {
-			throw new IllegalStateException("root path does not have a base property");
-		}
-		return path.getBaseProperty();
-	}
+	RelationalPersistentProperty getBaseProperty();
 
 	/**
 	 * The column name of the id column of the ancestor path that represents an actual table.
 	 */
-	public SqlIdentifier getIdColumnName() {
-		return getTableOwningAncestor().getRequiredLeafEntity().getIdColumn();
-	}
+	SqlIdentifier getIdColumnName();
 
 	/**
 	 * @return {@literal true} when this is references a {@link java.util.Collection} or an array.
 	 */
-	public boolean isCollectionLike() {
-		return !isRoot() && path.getLeafProperty().isCollectionLike();
-	}
+	boolean isCollectionLike();
 
 	/**
 	 * @return whether the leaf end of the path is ordered, i.e. the data to populate must be ordered.
 	 * @see RelationalPersistentProperty#isOrdered()
 	 */
-	public boolean isOrdered() {
-		return !isRoot() && path.getLeafProperty().isOrdered();
-	}
+	boolean isOrdered();
 
 	/**
 	 * Creates a new path by extending the current path by the property passed as an argument.
@@ -422,53 +196,29 @@ public class AggregatePath {
 	 * @param property must not be {@literal null}.
 	 * @return Guaranteed to be not {@literal null}.
 	 */
-	public AggregatePath append(RelationalPersistentProperty property) {
+	AggregatePath append(RelationalPersistentProperty property);
 
-		PersistentPropertyPath<? extends RelationalPersistentProperty> newPath = isRoot() //
-				? context.getPersistentPropertyPath(property.getName(), rootType.getType()) //
-				: context.getPersistentPropertyPath(path.toDotPath() + "." + property.getName(),
-						path.getBaseProperty().getOwner().getType());
-
-		return context.getAggregatePath(newPath);
-	}
-
-	public PersistentPropertyPath<? extends RelationalPersistentProperty> getRequiredPersistentPropertyPath() {
-
-		Assert.state(!isRoot(), "path must not be null");
-		return path;
-	}
+	PersistentPropertyPath<? extends RelationalPersistentProperty> getRequiredPersistentPropertyPath();
 
 	/**
 	 * If the table owning ancestor has an id the column name of that id property is returned. Otherwise the reverse
 	 * column is returned.
 	 */
-	public SqlIdentifier getEffectiveIdColumnName() {
-
-		AggregatePath owner = getTableOwningAncestor();
-		return owner.isRoot() ? owner.getRequiredLeafEntity().getIdColumn() : owner.getReverseColumnName();
-	}
+	SqlIdentifier getEffectiveIdColumnName();
 
 	@Nullable
-	public PersistentPropertyPathExtension getPathExtension() {
+	PersistentPropertyPathExtension getPathExtension();
 
-		if (isRoot()) {
-			return new PersistentPropertyPathExtension(context, rootType);
-		}
-		return new PersistentPropertyPathExtension(context, path);
-	}
+	/**
+	 * Finds and returns the longest path with ich identical or an ancestor to the current path and maps directly to a
+	 * table.
+	 *
+	 * @return a path. Guaranteed to be not {@literal null}.
+	 */
+	AggregatePath getTableOwningAncestor();
 
-	@Override
-	public boolean equals(Object o) {
+	@Nullable
+	SqlIdentifier assembleTableAlias();
 
-		if (this == o) return true;
-		if (o == null || getClass() != o.getClass()) return false;
-		AggregatePath that = (AggregatePath) o;
-		return Objects.equals(context, that.context) && Objects.equals(rootType, that.rootType) && Objects.equals(path, that.path);
-	}
-
-	@Override
-	public int hashCode() {
-
-		return Objects.hash(context, rootType, path);
-	}
+	SqlIdentifier assembleColumnName(SqlIdentifier suffix);
 }
